+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Added constructors and query parameter factory methods to request configuration classes and constructors to query parameter classes in PHP.
 
 ### Changed
 
 - Fixed a bug where go refiner would fail with a null reference.
+- Fixes a bug where request builders would be incorrectly generated due to inconsistent suffix generation
 - Fixed a bug where PHP model getAdditionalData() would not return nullable types.
 
 ## [0.11.1] - 2023-02-13